+++ CHANGELOG.md
@@ -46,6 +46,8 @@
 * *BREAKING* Instance: Display all versions in View history fourth pane. Refs UIIN-3173.
 * *BREAKING* MARC bib > View Source > Display Version History. Refs UIIN-3261.
 * "Copy barcode" icon is displayed next to item with no barcode. Fixes UIIN-3141.
+* *BREAKING* Holdings: Display all versions in View history second pane. Refs UIIN-3174.
+* *BREAKING* Item: Display all versions in View history second pane. Refs UIIN-3175.
 * Replace annotations for compatibility with esbuild-loader. Refs UIIN-3271.
 
 ## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)